// src/services/api/generated/azeel.schemas.ts
export type UsersStoreBodyStatus = (typeof UsersStoreBodyStatus)[keyof typeof UsersStoreBodyStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UsersStoreBodyStatus = {
  Active: "Active",
  Inactive: "Inactive",
} as const;